+++ src/commands/ProxyController.php
@@ -68,32 +68,33 @@ use luya\console\Command;
 class ProxyController extends Command
 {
     use CacheableTrait;
-    
+
     const CONFIG_VAR_URL = 'lcpProxyUrl';
-    
+
     const CONFIG_VAR_TOKEN = 'lcpProxyToken';
-    
+
     const CONFIG_VAR_IDENTIFIER = 'lcpProxyIdentifier';
-    
+
     /**
      * @inheritdoc
      */
     public $defaultAction = 'sync';
-    
+
     /**
      * @var boolean Whether the isComplet sync check should be done after finish or not. If a table has a lot of traffic sometimes
      * there is a difference between the exchange of table informations (build) and transfer the data. In order to prevent
      * the exception message you can disable the strict compare mode. In order to ensure strict comparing enable $strict.
      */
     public $strict = false;
-    
+
     /**
      * @var string If a table option is passed only this table will be synchronised. If false by default all tables will be synced. You
      * can define multible tables ab seperating those with a comma `table1,table2,table`. In order to define only tables with start
      * with a given prefix you can use `app_*` using asterisks symbold to define wild card starts with string defintions.
+     * To exclude tables you can use a `!` before the tablename e.g. `!admin_*` or multible `!admin_*,!test_*`
      */
     public $table;
-    
+
     /**
      * @var string The production environment Domain where your LUYA application is running in production mode make so to use the right protocolo
      * examples:
@@ -102,7 +103,7 @@ class ProxyController extends Command
      *
      */
     public $url;
-    
+
     /**
      * @var string The identifier you get from the Machines menu in your production env admin looks like this: lcp58e35acb4ca69
      */
@@ -117,7 +118,7 @@ class ProxyController extends Command
 	 * @var integer Number of requests collected until they are written to the database.
 	 */
     public $syncRequestsCount = 10;
-    
+
     /**
      * @inheritdoc
      */
@@ -125,7 +126,7 @@ class ProxyController extends Command
     {
         return array_merge(parent::options($actionID), ['strict', 'table', 'url', 'idf', 'token', 'syncRequestsCount']);
     }
-    
+
     /**
      * @inheritdoc
      */
@@ -133,7 +134,7 @@ class ProxyController extends Command
     {
         return array_merge(parent::optionAliases(), ['s' => 'strict', 't' => 'table', 'u' => 'url', 'i' => 'idf', 'tk' => 'token']);
     }
-    
+
     /**
      * Sync Proxy Data.
      *
@@ -143,7 +144,7 @@ class ProxyController extends Command
     {
         if ($this->url === null) {
             $url = Config::get(self::CONFIG_VAR_URL);
-            
+
             if (!$url) {
                 $url = $this->prompt('Enter the Proxy PROD env URL (e.g. https://example.com):');
                 Config::set(self::CONFIG_VAR_URL, $url);
@@ -151,10 +152,10 @@ class ProxyController extends Command
         } else {
             $url = $this->url;
         }
-        
+
         if ($this->idf === null) {
             $identifier = Config::get(self::CONFIG_VAR_IDENTIFIER);
-            
+
             if (!$identifier) {
                 $identifier = $this->prompt('Please enter the identifier:');
                 Config::set(self::CONFIG_VAR_IDENTIFIER, trim($identifier));
@@ -162,10 +163,10 @@ class ProxyController extends Command
         } else {
             $identifier = $this->idf;
         }
-        
+
         if ($this->token === null) {
             $token = Config::get(self::CONFIG_VAR_TOKEN);
-            
+
             if (!$token) {
                 $token = $this->prompt('Please enter the access token:');
                 Config::set(self::CONFIG_VAR_TOKEN, trim($token));
@@ -173,17 +174,17 @@ class ProxyController extends Command
         } else {
             $token = $this->token;
         }
-        
-        
+
+
         $proxyUrl = Url::ensureHttp(rtrim(trim($url), '/')) . '/admin/api-admin-proxy';
         $this->outputInfo('Connect to: ' . $proxyUrl);
-        
+
         $curl = new Curl();
         $curl->get($proxyUrl, ['identifier' => $identifier, 'token' => sha1($token)]);
-        
+
         if (!$curl->error) {
             $this->flushHasCache();
-            
+
             $this->verbosePrint($curl->response);
             $response = Json::decode($curl->response);
             $build = new ClientBuild($this, [
@@ -199,18 +200,18 @@ class ProxyController extends Command
                 'machineIdentifier' => $identifier,
                 'machineToken' => sha1($token),
             ]);
-            
+
             $process = new ClientTransfer(['build' => $build]);
             if ($process->start()) {
                 // as the admin_config table is synced to, we have to restore the current active config which has been used.
                 Config::set(self::CONFIG_VAR_IDENTIFIER, $identifier);
                 Config::set(self::CONFIG_VAR_TOKEN, $token);
                 Config::set(self::CONFIG_VAR_URL, $url);
-                
+
                 return $this->outputSuccess('Sync process has been successfully finished.');
             }
         }
-        
+
         $this->clearConfig();
         $this->output($curl->response);
         return $this->outputError($curl->error_message);
@@ -222,7 +223,7 @@ class ProxyController extends Command
         Config::remove(self::CONFIG_VAR_URL);
         Config::remove(self::CONFIG_VAR_IDENTIFIER);
     }
-    
+
     /**
      * Cleanup all stored Config Data.
      *