+++ src/ngrest/plugins/CmsPage.php
@@ -13,26 +13,42 @@ use luya\admin\ngrest\base\Plugin;
  */
 class CmsPage extends Plugin
 {
+    /**
+     * @inheritdoc
+     */
     public function renderList($id, $ngModel)
     {
-        return $this->createListTag($ngModel);
+        return $this->createTag('show-internal-redirection', null, ['nav-id' => $ngModel, 'ng-show' => $ngModel]);
     }
 
+    /**
+     * @inheritdoc
+     */
     public function renderCreate($id, $ngModel)
     {
         return $this->createFormTag('zaa-cms-page', $id, $ngModel);
     }
 
+    /**
+     * @inheritdoc
+     */
     public function renderUpdate($id, $ngModel)
     {
         return $this->renderCreate($id, $ngModel);
     }
 
+    /**
+     * @inheritdoc
+     */
     public function onAfterFind($event)
     {
+        // get value
         $fieldValue = $event->sender->getAttribute($this->name);
-        $menuItem = (!empty($fieldValue)) ? Yii::$app->menu->find()->where(['nav_id' => $fieldValue])->with(['hidden'])->one() : false;
         
+        // get menu item
+        $menuItem = !empty($fieldValue) ? Yii::$app->menu->find()->where(['nav_id' => $fieldValue])->with(['hidden'])->one() : false;
+        
+        // assign value
         $this->writeAttribute($event, $menuItem);
     }
 }